requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
